Add periodically updating RESTMapper
Currently, we fetch the discovery information once at the start of
of the adapter, and then never update it.  This could prove problematic,
since other API servers might come only after the adapter is started up.

Now, we periodically refresh the discovery information (as specified by
a flag), so that we have a chance to fetch new resources.